+++ CHANGELOG.md
@@ -17,6 +17,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Add conda path configuration in settings, default `conda` for global. If meet wrong like `conda: not found`, please set the path to conda executable in settings.
 
+### Remove
+
+- Remove npm test
+
 ## [2.0.2-beta] - 2026-02-24
 
 This version is a beta release for using python package [sage-lsp](https://pypi.org/project/sage-lsp/) to start LSP server. Same as [2.0.1]