+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "SageMath for VScode" extension will be documented in
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2026-xx-xx
+## [2.1.0] - 2026-04-09
 
 ### Added